add enemy pool and spawner

+++ Assets/[GameLogic]/Enemy/Components/RewardComponent.cs
@@ -5,6 +5,6 @@ namespace _GameLogic_.Enemy
     [Game]
     public sealed class RewardComponent : IComponent
     {
-        public float value;
+        public int value;
     }
 }
